blogs improvments

+++ views/src/app/main-page/blogs/blogs.component.ts
@@ -141,6 +141,7 @@ export class BlogsComponent implements OnInit {
     this.inManage = false;
   }
   delete() {
+    if (!confirm("Do you want delete these posts?\nAre you sure?")) return;
     this.selectedBlogs.forEach(blog => {
       this.appService.delete_blog(blog.id).subscribe(res => {
         if (res.status) {
